+++ firebase-vscode/common/messaging/protocol.ts
@@ -45,8 +45,6 @@ export interface WebviewToExtensionParamsMap {
   getEmulatorUiSelections: void;
   getEmulatorInfos: void;
   updateEmulatorUiSelections: Partial<EmulatorUiSelections>;
-  /* Equivalent to the `firebase emulators:start` command.*/
-  launchEmulators: void;
 
   /** Notify extension that current user has been changed in UI. */
   requestChangeUser: { user: User | ServiceAccountUser };
@@ -95,8 +93,9 @@ export interface WebviewToExtensionParamsMap {
     href: string;
   };
 
-  /** Stops the emulators gracefully allowing for data export if required. */
-  stopEmulators: void;
+  connectToPostgres: void;
+  disconnectPostgres: void;
+  getInitialIsConnectedToPostgres: void;
 
   selectEmulatorImportFolder: {};
 
@@ -139,6 +138,10 @@ export interface ExtensionToWebviewParamsMap {
     infos: RunningEmulatorInfo | undefined;
   };
   notifyEmulatorImportFolder: { folder: string };
+
+  notifyIsConnectedToPostgres: boolean;
+
+  notifyPostgresStringChanged: string;
 
   /** Triggered when new environment variables values are found. */
   notifyEnv: { env: { isMonospace: boolean } };