+++ firebase-vscode/webviews/data-connect.entry.tsx
@@ -1,11 +1,9 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
-import { VSCodeButton } from "@vscode/webview-ui-toolkit/react";
+import { VSCodeButton, VSCodeTextField } from "@vscode/webview-ui-toolkit/react";
 import { Spacer } from "./components/ui/Spacer";
 import styles from "./globals/index.scss";
-import { TEXT } from "./globals/ux-text";
-import { broker, useBroker } from "./globals/html-broker";
-import { Heading, Label } from "./components/ui/Text";
+import { broker, useBroker, useBrokerListener } from "./globals/html-broker";
 import { PanelSection } from "./components/ui/PanelSection";
 
 // Prevent webpack from removing the `style` import above
@@ -15,42 +13,41 @@ const root = createRoot(document.getElementById("root")!);
 root.render(<DataConnect />);
 
 function DataConnect() {
-  const emulatorsStatus =
-    useBroker("notifyEmulatorStateChanged", {
-      initialRequest: "getEmulatorInfos",
-    })?.status ?? "stopped";
+  const isConnectedToPostgres =
+    useBroker("notifyIsConnectedToPostgres", {
+      initialRequest: "getInitialIsConnectedToPostgres",
+    }) ?? false;
+
+  const psqlString = useBroker("notifyPostgresStringChanged");
 
   return (
     <>
-      <PanelSection title="Emulator">
-        <p>
-          Start the FDC emulator. See also:{" "}
-          <a href="https://firebase.google.com/docs/data-connect/quickstart">
-            Working with the emulator
-          </a>
-        </p>
+      <PanelSection title="Local Development">
+        {!isConnectedToPostgres && (<p>
+            Connect to Local PostgreSQL. See also:{" "}
+            <a href="https://firebase.google.com/docs/data-connect/quickstart">
+              Working with the emulator
+            </a>
+          </p>)
+        }
         <Spacer size="xsmall" />
-        {emulatorsStatus === "running" || emulatorsStatus === "stopping" ? (
-          <VSCodeButton
-            onClick={() => broker.send("stopEmulators")}
-            disabled={emulatorsStatus === "stopping"}
-          >
-            Stop emulator
-          </VSCodeButton>
+        {isConnectedToPostgres ? (
+          <>
+          <label>Local emulator connected to:</label>
+            <VSCodeTextField disabled value={psqlString}></VSCodeTextField>
+          </>
         ) : (
-          <VSCodeButton
-            onClick={() => broker.send("launchEmulators")}
-            disabled={emulatorsStatus === "starting"}
-          >
-            Start emulator
+          <VSCodeButton onClick={() => broker.send("connectToPostgres")}>
+            Connect to Local PostgreSQL
           </VSCodeButton>
         )}
       </PanelSection>
-
       <PanelSection title="Production" isLast={true}>
         <p>
           Deploy FDC services and connectors to production. See also:{" "}
-          <a href="https://firebase.google.com/docs/data-connect/quickstart">Deploying</a>
+          <a href="https://firebase.google.com/docs/data-connect/quickstart">
+            Deploying
+          </a>
         </p>
         <Spacer size="xsmall" />
         <VSCodeButton onClick={() => broker.send("fdc.deploy")}>