+++ src/emulator/controller.ts
@@ -843,6 +843,9 @@ export async function startAll(
       rc: options.rc,
     });
     await startEmulator(dataConnectEmulator);
+    if (!utils.isVSCodeExtension()) {
+      await dataConnectEmulator.connectToPostgres();
+    }
   }
 
   if (listenForEmulator.storage) {