+++ src/emulator/downloadableEmulators.ts
@@ -33,9 +33,9 @@ const EMULATOR_UPDATE_DETAILS: { [s in DownloadableEmulators]: EmulatorUpdateDet
     expectedChecksum: "2fd771101c0e1f7898c04c9204f2ce63",
   },
   firestore: {
-    version: "1.19.6",
-    expectedSize: 66349770,
-    expectedChecksum: "2eaabbe3cdb4867df585b7ec5505bad7",
+    version: "1.19.7",
+    expectedSize: 66438992,
+    expectedChecksum: "aec233bea95c5cfab03881574ec16d6c",
   },
   storage: {
     version: "1.1.3",
@@ -50,27 +50,27 @@ const EMULATOR_UPDATE_DETAILS: { [s in DownloadableEmulators]: EmulatorUpdateDet
         expectedChecksum: "a7f4398a00e5ca22abdcd78dc3877d00",
       },
   pubsub: {
-    version: "0.8.2",
-    expectedSize: 65611398,
-    expectedChecksum: "70bb840321423e6ae621a3ae2f314903",
+    version: "0.8.14",
+    expectedSize: 66786933,
+    expectedChecksum: "a9025b3e53fdeafd2969ccb3ba1e1d38",
   },
   dataconnect:
     process.platform === "darwin"
       ? {
-          version: "1.1.19",
-          expectedSize: 25836864,
-          expectedChecksum: "b31ba00789b82a30f0dee1c03b8f74ce",
+          version: "1.2.0",
+          expectedSize: 23954240,
+          expectedChecksum: "0f250761959519bb5a28fed76ceab2cb",
         }
       : process.platform === "win32"
         ? {
-            version: "1.1.19",
-            expectedSize: 23629824,
-            expectedChecksum: "bcbd7705b36cee72ff0587749d67bfc3",
+            version: "1.2.0",
+            expectedSize: 24360960,
+            expectedChecksum: "168ce32c742e1d26037c52bdbb7d871c",
           }
         : {
-            version: "1.1.19",
-            expectedSize: 23247120,
-            expectedChecksum: "56d6cb2ad85474d3a67999e35d2916a1",
+            version: "1.2.0",
+            expectedSize: 23970052,
+            expectedChecksum: "2ca17e4009a9ebae0f7c983bafff2ee6",
           },
 };
 
@@ -292,12 +292,15 @@ const Commands: { [s in DownloadableEmulators]: DownloadableEmulatorCommand } = 
     optionalArgs: [
       "listen",
       "config_dir",
-      "local_connection_string",
       "project_id",
       "service_location",
+      "disable_sdk_generation",
+      "resolvers_emulator",
+      "vertex_location",
+      "rpc_retry_count",
     ],
     joinArgs: true,
-    shell: true,
+    shell: false,
   },
 };
 
@@ -323,7 +326,6 @@ export function _getCommand(
   args: { [s: string]: any },
 ): DownloadableEmulatorCommand {
   const baseCmd = Commands[emulator];
-
   const defaultPort = Constants.getDefaultPort(emulator);
   if (!args.port) {
     args.port = defaultPort;
@@ -503,7 +505,21 @@ async function _runBinary(
  * @param emulator
  */
 export function getDownloadDetails(emulator: DownloadableEmulators): EmulatorDownloadDetails {
-  return DownloadDetails[emulator];
+  const details = DownloadDetails[emulator];
+  const pathOverride = process.env[`${emulator.toUpperCase()}_EMULATOR_BINARY_PATH`];
+  if (pathOverride) {
+    const logger = EmulatorLogger.forEmulator(emulator);
+    logger.logLabeled(
+      "WARN",
+      emulator,
+      `Env variable override detected. Using ${emulator} emulator at ${pathOverride}`,
+    );
+    details.downloadPath = pathOverride;
+    details.binaryPath = pathOverride;
+    details.localOnly = true;
+    fs.chmodSync(pathOverride, 0o755);
+  }
+  return details;
 }
 
 /**
@@ -579,7 +595,7 @@ export async function start(
   },
   extraEnv: Partial<NodeJS.ProcessEnv> = {},
 ): Promise<void> {
-  const downloadDetails = DownloadDetails[targetName];
+  const downloadDetails = getDownloadDetails(targetName);
   const emulator = get(targetName);
   const hasEmulator = fs.existsSync(getExecPath(targetName));
   const logger = EmulatorLogger.forEmulator(targetName);